+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
-- * 453fa04 - 2017-11-09: Release v9.1.69 
+- * 33de5b6 - 2017-11-10: Release v9.1.70 
+- * 9498449 - 2017-11-10: Fixed typo in code for daily searcher 
+- * b26010a - 2017-11-09: Release v9.1.69 
 - * e0c1882 - 2017-11-09: Fixed typo for server status template 
 - * fe471c0 - 2017-11-09: Fixed typo for server status template 
 - * 527cc77 - 2017-11-09: Updated ThePirateBay to work with most of the proxies out there